+ fixed VCF printing  + some static functions became inline  + removed warnings related to calloc  + solving problems with overlapping paired reads that do not match (the overlapping regions do not match)

+++ include/kmer-graph-types.h
@@ -68,7 +68,7 @@ typedef struct _KMER_EDGE_PAIR {
 	PKMER_EDGE U;
 	PKMER_EDGE V;
 	PKMER_EDGE ConnectingEdge;
-	uint32_t ReadDistance;
+	size_t ReadDistance;
 	PKMER_EDGE *Edges;
 	size_t EdgeCount;
 } KMER_EDGE_PAIR, *PKMER_EDGE_PAIR;
